Move GridSample shape inference into a grid_sample submodule
Move the GridSample operator, plus its test, from ops.rs into a new
ops/grid_sample.rs submodule, matching how this operator is grouped in
src/ops.rs.
